Refactor About section into modular components
- Created AboutBackground component for background decoration.
- Created AboutHeader component for section label and heading.
- Created AboutImage component for profile image and badges.
- Created AboutBio component for bio text and expertise.
- Created AboutHighlights component for highlight cards.
- Removed inline animations and data from About component for better readability and maintainability.
